+++ examples/mnist-mlp/Main.hs
@@ -14,8 +14,9 @@ import Torch.Functional hiding (take)
 import Torch.Tensor
 import Torch.NN
 import Torch.Optim
+import qualified Torch.Vision as I
 
-import qualified Image as I
+-- import qualified Image as I
 import qualified UntypedImage as UI
 
 data MLPSpec = MLPSpec {
