+++ hasktorch/src/Torch/Vision.hs
@@ -1,110 +1,68 @@
-{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE NoStarIsType #-}
-{-# LANGUAGE TypeOperators #-}
 
 module Torch.Vision where
 
-import           Control.Monad (forM_)
-import           System.IO.Unsafe
-import qualified Codec.Compression.GZip        as GZip
-import qualified Data.ByteString               as BS
-import qualified Data.ByteString.Lazy          as BS.Lazy
-import qualified Data.ByteString.Internal      as BSI
-import           GHC.TypeLits
+import Prelude hiding (min, max)
+import qualified Prelude as P
+
+import Control.Monad (forM_)
 
 import           Torch.Internal.Cast
-import           Torch.Typed.Aux
-import           Torch.Typed.Tensor
-import           Torch.Typed.Functional
-import qualified Torch.DType                   as D
+import qualified Data.ByteString as BS
+import qualified Data.ByteString.Internal as BSI
+import qualified Torch.DType as D
 import qualified Torch.Tensor                  as D
 import qualified Torch.TensorOptions           as D
 import qualified Foreign.ForeignPtr            as F
 import qualified Foreign.Ptr                   as F
 import qualified Torch.Internal.Managed.TensorFactories as LibTorch
 
-data MnistData =
-  MnistData
-  { images :: BS.ByteString
-  , labels :: BS.ByteString
-  }
+import qualified Torch.Typed.Vision as I
+import Torch.Functional hiding (take)
+import Torch.Tensor
+import Torch.NN
 
-type Rows = 28
-type Cols = 28
-type DataDim = Rows * Cols
-type ClassDim = 10
+getLabels' :: Int -> I.MnistData -> [Int] -> Tensor
+getLabels' n mnist imageIdxs =
+  asTensor $ map (I.getLabel mnist) . take n $ imageIdxs
 
-getLabels
-  :: forall n . KnownNat n => MnistData -> [Int] -> CPUTensor 'D.Int64 '[n]
-getLabels mnist imageIdxs =
-  UnsafeMkTensor . D.asTensor . map (getLabel mnist) . take (natValI @n) $ imageIdxs
-
-getLabel :: MnistData -> Int -> Int
-getLabel mnist imageIdx =
-  fromIntegral $ BS.index (labels mnist) (fromIntegral imageIdx + 8)
-
-getImage :: MnistData -> Int -> CPUTensor 'D.Float '[DataDim]
-getImage mnist imageIdx =
-  let imageBS =
-          [ fromIntegral $ BS.index (images mnist)
-                                    (fromIntegral imageIdx * 28 ^ 2 + 16 + r)
-          | r <- [0 .. 28 ^ 2 - 1]
-          ] :: [Float]
-      (tensor :: CPUTensor 'D.Float '[DataDim]) =
-          UnsafeMkTensor $ D.asTensor imageBS
-  in  tensor
-
-getImages'
-  :: forall n
-   . KnownNat n
-  => MnistData
-  -> [Int]
-  -> CPUTensor 'D.Float '[n, DataDim]
-getImages' mnist imageIdxs = UnsafeMkTensor $ D.asTensor $ map image $ take
-  (natValI @n)
-  imageIdxs
- where
-  image idx =
-    [ fromIntegral
-        $ BS.index (images mnist) (fromIntegral idx * 28 ^ 2 + 16 + r)
-    | r <- [0 .. 28 ^ 2 - 1]
-    ] :: [Float]
-
-getImages
-  :: forall n
-   . KnownNat n
-  => MnistData
-  -> [Int]
-  -> CPUTensor 'D.Float '[n, DataDim]
-getImages mnist imageIdxs = UnsafeMkTensor $ unsafePerformIO $ do
-  let (BSI.PS fptr off len) = images mnist
+getImages' ::
+  Int -- number of observations in minibatch
+  -> Int -- dimensionality of the data
+  -> I.MnistData -- mnist data representation
+  -> [Int] -- indices of the dataset
+  -> IO Tensor
+getImages' n dataDim mnist imageIdxs = do
+  let (BSI.PS fptr off len) = I.images mnist
   t <- (cast2 LibTorch.empty_lo :: [Int] -> D.TensorOptions -> IO D.Tensor)
-         [natValI @n, natValI @DataDim]
+         [n, dataDim]
          (D.withDType D.UInt8 D.defaultOpts)
   D.withTensor t $ \ptr1 -> do
     F.withForeignPtr fptr $ \ptr2 -> do
-      forM_ (zip [0..((natValI @n)-1)] imageIdxs) $ \(i,idx) -> do
+      forM_ (zip [0..(n-1)] imageIdxs) $ \(i,idx) -> do
         BSI.memcpy
-          (F.plusPtr ptr1 ((natValI @DataDim)*i))
-          (F.plusPtr ptr2 (off+16+(natValI @DataDim)*idx))
-          (natValI @DataDim)
+          (F.plusPtr ptr1 (dataDim*i))
+          (F.plusPtr ptr2 (off+16+dataDim*idx))
+          dataDim
   return $ D.toType D.Float t
 
-length :: MnistData -> Int
-length mnist = fromIntegral $ BS.length (labels mnist) - 8
+-- http://paulbourke.net/dataformats/asciiart/
+grayScale10 = " .:-=+*#%@"
+grayScale70 = reverse "$@B%8&WM#*oahkbdpqwmZO0QLCJUYXzcvunxrjft/\\|()1{}[]?-_+~<>i!lI;:,\"^`'. "
 
-decompressFile :: String -> String -> IO BS.ByteString
-decompressFile path file = decompress' <$> BS.readFile (path <> "/" <> file)
-  where
-    decompress' = BS.concat . BS.Lazy.toChunks . GZip.decompress . BS.Lazy.fromStrict
-
-initMnist :: IO (MnistData, MnistData)
-initMnist = do
-  let path = "data"
-  imagesBS <- decompressFile path "train-images-idx3-ubyte.gz"
-  labelsBS <- decompressFile path "train-labels-idx1-ubyte.gz"
-  testImagesBS <- decompressFile path "t10k-images-idx3-ubyte.gz"
-  testLabelsBS <- decompressFile path "t10k-labels-idx1-ubyte.gz"
-  return (MnistData imagesBS labelsBS, MnistData testImagesBS testLabelsBS)
+-- Display an MNIST image tensor as ascii text
+dispImage :: Tensor -> IO ()
+dispImage img = do
+    mapM (\row ->
+        mapM (\col -> 
+            putChar $ grayScale !! (P.floor $ scaled !! row !! col)
+            ) [0,downSamp..27] >>  putStrLn ""
+        ) [0,downSamp..27]
+    pure ()
+    where 
+        downSamp = 2
+        grayScale = grayScale10
+        paletteMax = (fromIntegral $ length grayScale) - 1.0
+        img' = reshape [28, 28] img
+        scaled :: [[Float]] = let (mn, mx) = (min img', max img')  
+            in asValue $ (img' - mn) / (mx - mn) * paletteMax
